fix: critical race conditions, speed scoring, thread safety, 14 bug fixes total

+++ src/main/java/com/sprintjudge/domain/dto/QuestionStart.java
@@ -3,7 +3,7 @@ package com.sprintjudge.domain.dto;
 public record QuestionStart(
         String type,
         QuestionDto question,
-        long timeLimitSec,
+        int timeLimitSec,
         long startedAtEpochMs,
         long serverNowEpochMs
 ) {}
